requirements.txt
# Todoist API
todoist-api-python==2.1.7

# LangChain & LangGraph
langchain>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-ollama>=1.0.0
langchain-anthropic>=1.0.0
langchain-openai>=1.0.0
langgraph>=1.0.0

# Streamlit for UI
streamlit==1.39.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.12.0

# Testing and development
jupyter==1.0.0
ipykernel==6.29.0
